fix: Remove env base
Both firmwares can be built at the same time with `pio run`, which is
convenient. By default, that command builds every environment. Since the
`base` environment has no platform, this results in a failed build for
the environment. There are two ways of eliminating the error:

1. Set the `default_envs` option in the `platformio` section:
https://docs.platformio.org/en/latest/projectconf/sections/platformio/options/generic/default_envs.html
2. Change `env:base` to just `env` so it defines the common environment
options without relying on the `extends` option:
https://docs.platformio.org/en/latest/projectconf/sections/env/index.html

Since only two environments are in use, and the `base` environment is
not intended to be built at all, this commit implements the second
option.